[CVD] Display all curves and fix some misbehaviors

+++ EYES/Assets/Scripts/EyeDiseases/Colorblindness/DiscreteFunction.cs
@@ -6,11 +6,11 @@ namespace eyediseases {
 
     public class DiscreteFunction {
 
-        public List<double> values = new List<double> ();
-        public double minX = 0.0;
-        public double maxX = 1.0;
+        public List<float> values = new List<float> ();
+        public float minX = 0.0f;
+        public float maxX = 1.0f;
 
-        public double Stepsize () {
+        public float Stepsize () {
             if (values.Count > 0) {
                 return (maxX - minX) / (values.Count - 1);
             } else {
@@ -38,7 +38,7 @@ namespace eyediseases {
 //            this.maxX = maxX;
 //        }
 
-        public static DiscreteFunction operator * (double c, DiscreteFunction f) {
+        public static DiscreteFunction operator * (float c, DiscreteFunction f) {
             DiscreteFunction g = new DiscreteFunction (f);
             for (int i = 0; i < g.values.Count; ++i) {
                 g.values[i] = c * f.values[i];
